Remove BlockUtil move logic to J.Block
