+++ rewrite-java/src/main/java/org/openrewrite/java/controlflow/ControlFlow.java
@@ -23,7 +23,6 @@ import org.openrewrite.Tree;
 import org.openrewrite.internal.lang.Nullable;
 import org.openrewrite.java.JavaIsoVisitor;
 import org.openrewrite.java.JavaTemplate;
-import org.openrewrite.java.internal.BlockUtil;
 import org.openrewrite.java.tree.*;
 
 import java.util.*;
@@ -62,7 +61,7 @@ public final class ControlFlow {
             Object next = nextCursor.getValue();
             if (next instanceof J.Block) {
                 methodDeclarationBlockCursor = nextCursor;
-                if (BlockUtil.isStaticOrInitBlock(nextCursor)) {
+                if (J.Block.isStaticOrInitBlock(nextCursor)) {
                     break;
                 }
             } else if (next instanceof J.MethodDeclaration) {
